BO: Separate translation visibility and edition
